auth performance ok
